+++ index.html
@@ -22,6 +22,7 @@
         <div id="modal-movie-info"></div>
       </div>
     </div>
+    <script type="module" src="storemovies.js"></script>
     <script src="movie.js"></script>
   </body>
 </html>
